+++ app/grails-app/conf/DataSource.groovy
@@ -6,6 +6,7 @@ dataSource {
 }
 
 hibernate { // to hibernate 4
+    default_schema = "public"
     cache.use_second_level_cache = true
     cache.use_query_cache = false  // LEGACY
     //cache.region.factory_class = 'org.hibernate.cache.ehcache.SingletonEhCacheRegionFactory' // hibernate4: CAUTION: USE FOR DB-MIGRATION-PLUGIN
@@ -17,11 +18,11 @@ environments {
     development {
         dataSource {
             dbCreate = "update"
-            driverClassName = "com.mysql.jdbc.Driver"
-            dialect = "org.hibernate.dialect.MySQL5Dialect"
+            driverClassName = "org.postgresql.Driver"
+            dialect = "org.hibernate.dialect.PostgreSQLDialect"
             username = "laser"
             password = "laser"
-            url = "jdbc:mysql://localhost/laser?autoReconnect=true&useUnicode=true&characterEncoding=UTF-8"
+            url = "jdbc:postgresql://localhost:5432/laser"
             pooled = true
             properties {
                 maxActive = -1
@@ -38,11 +39,11 @@ environments {
     test {
         dataSource {
             dbCreate = "create-drop"
-            driverClassName = "com.mysql.jdbc.Driver"
-            dialect = "org.hibernate.dialect.MySQL5Dialect"
+            driverClassName = "org.postgresql.Driver"
+            dialect = "org.hibernate.dialect.PostgreSQLDialect"
             username = "laser"
             password = "laser"
-            url = "jdbc:mysql://localhost/laserTest?autoReconnect=true&useUnicode=true&characterEncoding=UTF-8"
+            url = "jdbc:postgresql://localhost:5432/laser"
             pooled = true
             properties {
                 maxActive = -1
@@ -59,12 +60,11 @@ environments {
     production {
         dataSource {
             dbCreate = "none" // disabled due database migration plugin; overwritten on dev-server
-            driverClassName = "com.mysql.jdbc.Driver"
+            driverClassName = "org.postgresql.Driver"
+            dialect = "org.hibernate.dialect.PostgreSQLDialect"
             username = "laser"
             password = "laser"
-            url = "jdbc:mysql://localhost/laser?autoReconnect=true&useUnicode=true&characterEncoding=UTF-8"
-            pooled = true
-            dialect = "org.hibernate.dialect.MySQL5Dialect"
+            url = "jdbc:postgresql://localhost:5432/laser"
             properties {
                 maxActive = -1
                 minEvictableIdleTimeMillis=1800000