+++ app/grails-app/controllers/com/k_int/kbplus/GlobalDataSyncController.groovy
@@ -56,7 +56,7 @@ class GlobalDataSyncController extends AbstractDebugController {
       }
     }
 
-    result.globalItemTotal = Subscription.executeQuery("select count(r) " + base_qry, qry_params)[0]
+    result.globalItemTotal = Subscription.executeQuery("select r.id " + base_qry, qry_params).size()
     result.items = Subscription.executeQuery("select r ${base_qry}", qry_params, [max: result.max, offset: result.offset]);
 
     result.tippcount = []