+++ app/grails-app/controllers/com/k_int/kbplus/SubscriptionDetailsController.groovy
@@ -186,7 +186,7 @@ class SubscriptionDetailsController extends AbstractDebugController {
             base_qry += "order by lower(ie.tipp.title.title) asc"
         }
 
-        result.num_sub_rows = IssueEntitlement.executeQuery("select count(ie) " + base_qry, qry_params)[0]
+        result.num_sub_rows = IssueEntitlement.executeQuery("select ie.id " + base_qry, qry_params).size()
 
         if (params.format == 'html' || params.format == null) {
             result.entitlements = IssueEntitlement.executeQuery("select ie " + base_qry, qry_params, [max: result.max, offset: result.offset]);
@@ -1321,7 +1321,7 @@ AND l.status.value != 'Deleted' order by LOWER(l.reference)
                 qry = """
 select l from License as l 
 where exists ( select ol from OrgRole as ol where ol.lic = l AND ol.org = ? and ( ol.roleType = ?) ) 
-AND l.status.value != 'Deleted' AND (l.instanceOf is null or l.instanceOf = '') order by LOWER(l.reference)
+AND l.status.value != 'Deleted' AND (l.instanceOf is null) order by LOWER(l.reference)
 """
             }