+++ app/grails-app/controllers/com/k_int/kbplus/YodaController.groovy
@@ -104,9 +104,10 @@ class YodaController {
         def result = [:]
 
         result.byUri =
-                SystemProfiler.executeQuery("select sp, avg(sp.ms) as ms, count(*) from SystemProfiler sp group by sp.uri").sort{it[1]}.reverse()
+                //SystemProfiler.executeQuery("select sp, avg(sp.ms) as ms, count(*), sp.id from SystemProfiler sp group by sp.uri").sort{it[1]}.reverse()
+                SystemProfiler.executeQuery("select sp.uri, avg(sp.ms) as ms, count(sp.id) as count from SystemProfiler sp group by sp.uri").sort{it[1]}.reverse()
         result.byUriAndContext =
-                SystemProfiler.executeQuery("select sp, avg(sp.ms) as ms, count(*) from SystemProfiler sp group by sp.uri, sp.context").sort{it[1]}.reverse()
+                SystemProfiler.executeQuery("select sp.uri, org.id, avg(sp.ms) as ms, count(org.id) as count from SystemProfiler sp join sp.context as org group by sp.uri, org.id").sort{it[2]}.reverse()
 
         result
     }