+++ app/grails-app/domain/com/k_int/kbplus/Subscription.groovy
@@ -58,7 +58,6 @@ class Subscription extends AbstractBaseDomain implements TemplateSupport, Permis
   License owner
   SortedSet issueEntitlements
   RefdataValue isPublic     // RefdataCategory 'YN'
-  Set ids = []
 
   static transients = [ 'subscriber', 'provider', 'consortia' ]
 
@@ -87,6 +86,7 @@ class Subscription extends AbstractBaseDomain implements TemplateSupport, Permis
                       derivedSubscriptions: 'instanceOf',
                       pendingChanges: 'subscription',
                       costItems: 'sub',
+                      customProperties: 'owner',
                       privateProperties: 'owner',
                       oapl: 'subscription'
                       ]
@@ -133,8 +133,8 @@ class Subscription extends AbstractBaseDomain implements TemplateSupport, Permis
         isSlaved(nullable:true, blank:false)
         noticePeriod(nullable:true, blank:true)
         isPublic(nullable:true, blank:true)
-        customProperties(nullable:true)
-        privateProperties(nullable:true)
+        //customProperties(nullable:true)
+        //privateProperties(nullable:true)
         cancellationAllowances(nullable:true, blank:true)
         lastUpdated(nullable: true, blank: true)
         // vendor(nullable:true, blank:false)
@@ -534,7 +534,8 @@ class Subscription extends AbstractBaseDomain implements TemplateSupport, Permis
   }
 
   def setInstitution(inst) {
-    println("Set institution ${inst}");
+      log.debug("Set institution ${inst}")
+
     def subrole = RDStore.OR_SUBSCRIBER
     def or = new OrgRole(org:inst, roleType:subrole, sub:this)
     if ( this.orgRelations == null)
@@ -543,7 +544,8 @@ class Subscription extends AbstractBaseDomain implements TemplateSupport, Permis
   }
 
   def addNamespacedIdentifier(ns,value) {
-    println("Add Namespaced identifier ${ns}:${value}");
+      log.debug("Add Namespaced identifier ${ns}:${value}")
+
     def canonical_id = Identifier.lookupOrCreateCanonicalIdentifier(ns, value);
     if ( this.ids == null)
       this.ids = []