+++ app/grails-app/services/de/laser/PropertyService.groovy
@@ -87,7 +87,7 @@ class PropertyService {
                 //log.debug(query)
 
                 def pds = SystemAdmin.executeQuery(query)
-                println pds
+                log.debug(pds)
                 detailsMap << ["${dc.shortName}": pds.collect{ it -> "${it.id}:${it.type}:${it.descr}"}.sort()]
 
                 // ids of used property definitions